requirements.txt
colorama
pyfiglet
speedtest-cli